feat: improve styling and rename parent folder
